Se refactoriza las clases separandolas por agregados
