+++ back-springboot/infrastructure/entry-points/api-rest/src/main/java/co/com/sofka/api/count/CountDTO.java
@@ -1,4 +1,4 @@
-package co.com.sofka.api;
+package co.com.sofka.api.count;
 
 import co.com.sofka.model.count.values.IdCount;
 import co.com.sofka.model.count.values.Name;
